Moved process mark to own file
